requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
Brotli==1.1.0
build==1.2.1
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
cssselect2==0.7.0
Django==5.0.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-formtools==2.5.1
django-otp==1.5.0
django-phonenumber-field==7.3.0
django-timezone-field==6.1.0
django-two-factor-auth==1.16.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
fonttools==4.53.1
frozenlist==1.4.1
html5lib==1.1
idna==3.7
inflection==0.5.1
kombu==5.3.7
multidict==6.0.5
mysqlclient==2.2.4
packaging==24.1
phonenumbers==8.13.40
pillow==10.3.0
pip-tools==7.4.1
prompt_toolkit==3.0.47
pycparser==2.22
pydyf==0.11.0
PyJWT==2.8.0
pyphen==0.15.0
pypng==0.20220715.0
pyproject_hooks==1.1.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.6
requests==2.32.3
setuptools==70.3.0
six==1.16.0
sqlparse==0.5.0
tinycss2==1.3.0
twilio==9.2.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
weasyprint==62.3
webencodings==0.5.1
wheel==0.43.0
yarl==1.9.4
zopfli==0.2.3
